Differentiate Tutorials and Guides
These changes address the fact that it wasn't initially obvious in navigation what one would find in Guides and how that would differ from what one finds in Tutorials.
Optional: I'd suggest changing "Guides" to "Quick Guides" to emphasize this difference. If this change seems unnecessary/unhelpful, I'm okay with that! If we like it, we may need to update links in other places.