Refactor XForm widgets (#862)
The widgets were previously in a bulleted list, which made the flow difficult to read, and impossible to link to individual widgets in documentation. This version moves each widget to its own subsection so that they can be referenced properly. This structure also allows for further details to be included, such as screenshots and other details for their use. 

For https://github.com/medic/cht-docs/issues/861