Add "code suggestion" recommendation to review docs
For now this seems to be the best place for this recommendation, but ideally we have a set of more detailed pages on code reviews. We currently link out to other recommendations (cool, since there's no need to duplicate all that), but it means not really having a good landing page for all this. For the time being I'm adding this here, but if we expand our code review guidance, this section should move to whatever comes in the future.