Adding dark-mode favicon capability and images
Resolves #924

Our current theme inherits from docsy and the favicons are generates from the following template:
https://github.com/google/docsy/blob/1d126ab1bbf186fa0ea52a6d6a3f58229b2b7942/layouts/partials/favicons.html

This changes adds a new favicon layout to override the base. The new layout is the same as the base, but adds the media attribute to supply browsers with light/dark options should they choose to respect it.

White color images are included for all sizes for the dark option.