Initial simple GitHub codeowners setup
The goal here is to align parts of the handbook with people on the team who should be paying special attention. Different changes need different levels of attention.

This first file sets up a small group at the top level of the docs and the has a level set up where access rights are expanded to product management.

As we add more content to the handbook it is expected that new GitHub groups would be set up and those groups referneced here and aligned with specific areas of documentation.