Add product team core competencies
These were developed together as a team. We spoke about them in person in the 2023 Zanzibar meetup where everyone brought ideas, they were grouped and condensed, and put into a simple readable form, which can be seen here.

I'm adding these to the docs site, but the important point is that it wasn't just me authoring them.