fix(#1403): Improve 'Adding Privacy Policies to Apps'
Clarified location of privacy-policies.json file in the documentation

The guide has been updated to specify that the privacy-policies.json file should reside in the root of the project directory rather than inside the privacy_policies folder. This change ensures users place the file correctly to avoid configuration issues during the upload-privacy-policies action.